Extract panorama detection to Core
Move the pure panorama-detection logic (DetectPanoramas, elevation-tier
clustering, mosaic/panorama validation, quality tiering, and the
detection constants) out of the API's PanoramaService into a new
PanoramaDetector in MarsVista.Core. The logic depends only on Core
entities and ILogger, so both the API (request-time detection) and the
upcoming scraper pre-compute step can share one implementation.

Pure refactor: PanoramaService now delegates to an injected detector and
the DTO mapping stays in the API. All 65 existing panorama tests pass
unmodified.
