+++ tests/MarsVista.Api.Tests/Services/V2/PanoramaServiceTests.cs
@@ -22,6 +22,7 @@ public class PanoramaServiceTests : IntegrationTestBase
 
         services.AddSingleton(_mockLogger.Object);
         services.AddSingleton(_mockPhotoService.Object);
+        services.AddScoped<MarsVista.Core.Services.PanoramaDetector>();
         services.AddScoped<PanoramaService>();
     }
